Create Src.java
Main function that you run in main class. Always have main function and String must always be capitalized when declaring a variable. The rest of the stuff will be in the side classes. I will only be creating one object from another class for example, the rest of the classes will not be used.